+++ src/eval/symbols.js
@@ -107,7 +107,7 @@ function children(expr) {
     case 'Namespace':
       return e.args;
     case 'Member':
-      return [e.receiver];
+      return e.computed ? [e.receiver, e.key] : [e.receiver];
     case 'ArrayLit':
       return e.elements;
     case 'ObjectLit':